# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Creator Tools API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TextResponse {
  result: string;
}

export interface TextListResponse {
  results: string[];
}

export interface SimpleTextRequest {
  text: string;
}

export type YoutubeTitlesRequestStyle =
  (typeof YoutubeTitlesRequestStyle)[keyof typeof YoutubeTitlesRequestStyle];

export const YoutubeTitlesRequestStyle = {
  educational: "educational",
  entertaining: "entertaining",
  clickbait: "clickbait",
  listicle: "listicle",
  "how-to": "how-to",
} as const;

export interface YoutubeTitlesRequest {
  topic: string;
  style?: YoutubeTitlesRequestStyle;
}

export type InstagramBioRequestTone =
  (typeof InstagramBioRequestTone)[keyof typeof InstagramBioRequestTone];

export const InstagramBioRequestTone = {
  professional: "professional",
  funny: "funny",
  inspiring: "inspiring",
  casual: "casual",
} as const;

export interface InstagramBioRequest {
  name: string;
  niche: string;
  tone?: InstagramBioRequestTone;
}

export type HashtagsRequestPlatform =
  (typeof HashtagsRequestPlatform)[keyof typeof HashtagsRequestPlatform];

export const HashtagsRequestPlatform = {
  instagram: "instagram",
  twitter: "twitter",
  tiktok: "tiktok",
  linkedin: "linkedin",
} as const;

export interface HashtagsRequest {
  topic: string;
  platform: HashtagsRequestPlatform;
  /**
   * @minimum 5
   * @maximum 30
   */
  count?: number;
}

export type CaptionRequestPlatform =
  (typeof CaptionRequestPlatform)[keyof typeof CaptionRequestPlatform];

export const CaptionRequestPlatform = {
  instagram: "instagram",
  facebook: "facebook",
  linkedin: "linkedin",
  twitter: "twitter",
} as const;

export type CaptionRequestTone =
  (typeof CaptionRequestTone)[keyof typeof CaptionRequestTone];

export const CaptionRequestTone = {
  funny: "funny",
  professional: "professional",
  inspirational: "inspirational",
  casual: "casual",
} as const;

export interface CaptionRequest {
  description: string;
  platform: CaptionRequestPlatform;
  tone?: CaptionRequestTone;
}

export interface BlogIdeasRequest {
  niche: string;
  /**
   * @minimum 3
   * @maximum 10
   */
  count?: number;
}

export interface EmailSubjectRequest {
  emailTopic: string;
  audience?: string;
}

export type TweetRequestTone =
  (typeof TweetRequestTone)[keyof typeof TweetRequestTone];

export const TweetRequestTone = {
  informative: "informative",
  funny: "funny",
  controversial: "controversial",
  inspirational: "inspirational",
} as const;

export interface TweetRequest {
  topic: string;
  tone?: TweetRequestTone;
  /**
   * @minimum 1
   * @maximum 5
   */
  count?: number;
}

export interface ProductDescriptionRequest {
  productName: string;
  features: string;
  targetAudience?: string;
}

export type AdCopyRequestPlatform =
  (typeof AdCopyRequestPlatform)[keyof typeof AdCopyRequestPlatform];

export const AdCopyRequestPlatform = {
  google: "google",
  facebook: "facebook",
  instagram: "instagram",
  twitter: "twitter",
} as const;

export type AdCopyRequestGoal =
  (typeof AdCopyRequestGoal)[keyof typeof AdCopyRequestGoal];

export const AdCopyRequestGoal = {
  awareness: "awareness",
  clicks: "clicks",
  conversions: "conversions",
  engagement: "engagement",
} as const;

export interface AdCopyRequest {
  product: string;
  platform: AdCopyRequestPlatform;
  goal?: AdCopyRequestGoal;
}

export type ParaphraseRequestTone =
  (typeof ParaphraseRequestTone)[keyof typeof ParaphraseRequestTone];

export const ParaphraseRequestTone = {
  formal: "formal",
  casual: "casual",
  simple: "simple",
  creative: "creative",
} as const;

export interface ParaphraseRequest {
  text: string;
  tone?: ParaphraseRequestTone;
}

export interface GrammarCheckResponse {
  corrected: string;
  changes: string[];
}

export type SummarizeRequestLength =
  (typeof SummarizeRequestLength)[keyof typeof SummarizeRequestLength];

export const SummarizeRequestLength = {
  short: "short",
  medium: "medium",
  long: "long",
} as const;

export interface SummarizeRequest {
  text: string;
  length?: SummarizeRequestLength;
}

export interface SeoMetaRequest {
  pageTitle: string;
  description: string;
  keywords?: string;
}

export interface SeoMetaResponse {
  metaTitle: string;
  metaDescription: string;
  keywords: string[];
}