# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Creator Tools API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Generate YouTube titles
 */
export const GenerateYoutubeTitlesBody = zod.object({
  topic: zod.string(),
  style: zod
    .enum(["educational", "entertaining", "clickbait", "listicle", "how-to"])
    .optional(),
});

export const GenerateYoutubeTitlesResponse = zod.object({
  results: zod.array(zod.string()),
});

/**
 * @summary Generate Instagram bio
 */
export const GenerateInstagramBioBody = zod.object({
  name: zod.string(),
  niche: zod.string(),
  tone: zod.enum(["professional", "funny", "inspiring", "casual"]).optional(),
});

export const GenerateInstagramBioResponse = zod.object({
  result: zod.string(),
});

/**
 * @summary Generate hashtags
 */
export const generateHashtagsBodyCountMin = 5;
export const generateHashtagsBodyCountMax = 30;

export const GenerateHashtagsBody = zod.object({
  topic: zod.string(),
  platform: zod.enum(["instagram", "twitter", "tiktok", "linkedin"]),
  count: zod
    .number()
    .min(generateHashtagsBodyCountMin)
    .max(generateHashtagsBodyCountMax)
    .optional(),
});

export const GenerateHashtagsResponse = zod.object({
  results: zod.array(zod.string()),
});

/**
 * @summary Generate social media caption
 */
export const GenerateCaptionBody = zod.object({
  description: zod.string(),
  platform: zod.enum(["instagram", "facebook", "linkedin", "twitter"]),
  tone: zod
    .enum(["funny", "professional", "inspirational", "casual"])
    .optional(),
});

export const GenerateCaptionResponse = zod.object({
  result: zod.string(),
});

/**
 * @summary Generate blog post ideas
 */
export const generateBlogIdeasBodyCountMin = 3;
export const generateBlogIdeasBodyCountMax = 10;

export const GenerateBlogIdeasBody = zod.object({
  niche: zod.string(),
  count: zod
    .number()
    .min(generateBlogIdeasBodyCountMin)
    .max(generateBlogIdeasBodyCountMax)
    .optional(),
});

export const GenerateBlogIdeasResponse = zod.object({
  results: zod.array(zod.string()),
});

/**
 * @summary Generate email subject lines
 */
export const GenerateEmailSubjectBody = zod.object({
  emailTopic: zod.string(),
  audience: zod.string().optional(),
});

export const GenerateEmailSubjectResponse = zod.object({
  results: zod.array(zod.string()),
});

/**
 * @summary Generate tweets/X posts
 */
export const generateTweetBodyCountMax = 5;

export const GenerateTweetBody = zod.object({
  topic: zod.string(),
  tone: zod
    .enum(["informative", "funny", "controversial", "inspirational"])
    .optional(),
  count: zod.number().min(1).max(generateTweetBodyCountMax).optional(),
});

export const GenerateTweetResponse = zod.object({
  results: zod.array(zod.string()),
});

/**
 * @summary Generate product description
 */
export const GenerateProductDescriptionBody = zod.object({
  productName: zod.string(),
  features: zod.string(),
  targetAudience: zod.string().optional(),
});

export const GenerateProductDescriptionResponse = zod.object({
  result: zod.string(),
});

/**
 * @summary Generate ad copy
 */
export const GenerateAdCopyBody = zod.object({
  product: zod.string(),
  platform: zod.enum(["google", "facebook", "instagram", "twitter"]),
  goal: zod
    .enum(["awareness", "clicks", "conversions", "engagement"])
    .optional(),
});

export const GenerateAdCopyResponse = zod.object({
  results: zod.array(zod.string()),
});

/**
 * @summary Convert text to emoji
 */
export const TextToEmojiBody = zod.object({
  text: zod.string(),
});

export const TextToEmojiResponse = zod.object({
  result: zod.string(),
});

/**
 * @summary Paraphrase/rewrite text
 */
export const ParaphraseTextBody = zod.object({
  text: zod.string(),
  tone: zod.enum(["formal", "casual", "simple", "creative"]).optional(),
});

export const ParaphraseTextResponse = zod.object({
  result: zod.string(),
});

/**
 * @summary Grammar and spell check
 */
export const GrammarCheckBody = zod.object({
  text: zod.string(),
});

export const GrammarCheckResponse = zod.object({
  corrected: zod.string(),
  changes: zod.array(zod.string()),
});

/**
 * @summary Summarize text
 */
export const SummarizeTextBody = zod.object({
  text: zod.string(),
  length: zod.enum(["short", "medium", "long"]).optional(),
});

export const SummarizeTextResponse = zod.object({
  result: zod.string(),
});

/**
 * @summary Generate SEO meta title and description
 */
export const GenerateSeoMetaBody = zod.object({
  pageTitle: zod.string(),
  description: zod.string(),
  keywords: zod.string().optional(),
});

export const GenerateSeoMetaResponse = zod.object({
  metaTitle: zod.string(),
  metaDescription: zod.string(),
  keywords: zod.array(zod.string()),
});