# lib/api-zod/src/generated/types/adCopyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Creator Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { AdCopyRequestGoal } from "./adCopyRequestGoal";
import type { AdCopyRequestPlatform } from "./adCopyRequestPlatform";

export interface AdCopyRequest {
  product: string;
  platform: AdCopyRequestPlatform;
  goal?: AdCopyRequestGoal;
}