# lib/api-zod/src/generated/types/blogIdeasRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Creator Tools API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogIdeasRequest {
  niche: string;
  /**
   * @minimum 3
   * @maximum 10
   */
  count?: number;
}