# lib/api-zod/src/generated/types/captionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Creator Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { CaptionRequestPlatform } from "./captionRequestPlatform";
import type { CaptionRequestTone } from "./captionRequestTone";

export interface CaptionRequest {
  description: string;
  platform: CaptionRequestPlatform;
  tone?: CaptionRequestTone;
}