# lib/api-zod/src/generated/types/grammarCheckResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Creator Tools API
 * OpenAPI spec version: 0.1.0
 */

export interface GrammarCheckResponse {
  corrected: string;
  changes: string[];
}