# lib/api-zod/src/generated/types/hashtagsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Creator Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { HashtagsRequestPlatform } from "./hashtagsRequestPlatform";

export interface HashtagsRequest {
  topic: string;
  platform: HashtagsRequestPlatform;
  /**
   * @minimum 5
   * @maximum 30
   */
  count?: number;
}