# lib/api-zod/src/generated/types/instagramBioRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Creator Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { InstagramBioRequestTone } from "./instagramBioRequestTone";

export interface InstagramBioRequest {
  name: string;
  niche: string;
  tone?: InstagramBioRequestTone;
}