# lib/api-zod/src/generated/types/paraphraseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Creator Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { ParaphraseRequestTone } from "./paraphraseRequestTone";

export interface ParaphraseRequest {
  text: string;
  tone?: ParaphraseRequestTone;
}