# lib/api-zod/src/generated/types/productDescriptionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Creator Tools API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductDescriptionRequest {
  productName: string;
  features: string;
  targetAudience?: string;
}