# lib/api-zod/src/generated/types/summarizeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Creator Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { SummarizeRequestLength } from "./summarizeRequestLength";

export interface SummarizeRequest {
  text: string;
  length?: SummarizeRequestLength;
}