# lib/api-zod/src/generated/types/tweetRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Creator Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { TweetRequestTone } from "./tweetRequestTone";

export interface TweetRequest {
  topic: string;
  tone?: TweetRequestTone;
  /**
   * @minimum 1
   * @maximum 5
   */
  count?: number;
}