# lib/api-zod/src/generated/types/youtubeTitlesRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Creator Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { YoutubeTitlesRequestStyle } from "./youtubeTitlesRequestStyle";

export interface YoutubeTitlesRequest {
  topic: string;
  style?: YoutubeTitlesRequestStyle;
}